requirements-txt.txt
torch>=1.8.0
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
flwr>=0.17.0
pqcrypto>=0.1.3
pandas>=1.3.0
scikit-learn>=0.24.0
web3>=5.23.0
numpy>=1.21.0
matplotlib>=3.4.0
pyyaml>=5.4.0
